Refactor Snapshot service
This commit splits the snapshot service into several more granular components:

* SnapshotsService remains the faced class for all snapshot-related operations,
* SnapshotManager is now responsible for starting, deleting and aborting individual snapshots,
* SnapshotsShardService handles shard-level snapshot operations and is responsible for start and stopping snapshots of individual shards,
* SnapshotsShardWatcherService is used to monitor cluster state for the changes that affect the running snapshots
